+++ src/main/java/com/example/daedongv3_5/domain/club/presentation/dto/response/ClubResponse.java
@@ -1,15 +1,27 @@
 package com.example.daedongv3_5.domain.club.presentation.dto.response;
 
-import com.example.daedongv3_5.domain.club.domain.enums.Major;
+import com.example.daedongv3_5.domain.club.domain.Club;
+import com.example.daedongv3_5.domain.club.domain.Major;
+import com.example.daedongv3_5.domain.club.domain.enums.MajorType;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
+
+import java.util.List;
 
 @Getter
 @AllArgsConstructor
 public class ClubResponse {
-    private Long clubId;
+    private Long id;
     private String clubName;
     private String introduction;
     private String clubMember;
-    private Major major;
+    private List<Major> majors;
+
+    public ClubResponse(Club club) {
+        this.id = club.getId();
+        this.clubName = club.getClubName();
+        this.introduction = club.getIntroduction();
+        this.clubMember = club.getClubMember();
+        this.majors = club.getMajors();
+    }
 }
